+++ src/middlewares/context.middleware.ts
@@ -1,0 +1,35 @@
+import type {
+  ParamsContext,
+  RecorderState,
+  ServiceContext,
+  IOClients,
+} from '@vtex/api'
+import { prepareHandlerCtx } from '@vtex/api/lib/service/worker/runtime/utils/context'
+
+export const createContextMiddleware = () => {
+  return async function pvtContext<
+    T extends IOClients,
+    U extends RecorderState,
+    V extends ParamsContext
+  >(ctx: ServiceContext<T, U, V>, next: () => Promise<void>) {
+    const {
+      params,
+      request: { header },
+    } = ctx
+
+    // TODO get routeId from request
+    const routeId = ''
+
+    ctx.vtex = {
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      ...prepareHandlerCtx(header, ctx.tracing!),
+      // ...(smartcache && { recorder: ctx.state.recorder }),
+      route: {
+        id: routeId,
+        params,
+        type: 'private',
+      },
+    }
+    await next()
+  }
+}
